+++ src/client_core.d
@@ -1,2 +1,3 @@
-src/client_core.o: src/client_core.c src/client_core.h
+src/client_core.o: src/client_core.c src/client_core.h src/svc.h
 src/client_core.h:
+src/svc.h: