+++ src/main.d
@@ -11,8 +11,8 @@ src/main.o: src/main.c vendor/ffmpeg/include/libavutil/avutil.h \
  vendor/ffmpeg/include/libavutil/intfloat.h \
  vendor/ffmpeg/include/libavutil/log.h \
  vendor/ffmpeg/include/libavutil/pixfmt.h src/player.h src/client_core.h \
- src/stream_player.h src/plugin.h src/plugin_loader.h src/lang.h \
- src/update.h src/config.h src/cd.h
+ src/stream_player.h src/svc.h src/plugin.h src/plugin_loader.h \
+ src/lang.h src/update.h src/config.h src/cd.h
 vendor/ffmpeg/include/libavutil/avutil.h:
 vendor/ffmpeg/include/libavutil/common.h:
 vendor/ffmpeg/include/libavutil/attributes.h:
@@ -29,6 +29,7 @@ vendor/ffmpeg/include/libavutil/pixfmt.h:
 src/player.h:
 src/client_core.h:
 src/stream_player.h:
+src/svc.h:
 src/plugin.h:
 src/plugin_loader.h:
 src/lang.h: