+++ src/svc.d
@@ -1,0 +1,2 @@
+src/svc.o: src/svc.c src/svc.h
+src/svc.h: